BUILD_TYPE

+++ mango/mango.h
@@ -13,7 +13,7 @@ namespace mango {
 typedef unsigned int uint128_t __attribute__((mode(TI)));
 
 constexpr uint16_t safe_sub(uint16_t a, uint16_t b) {
-  return (a < b) ? 0 : a - b;
+  return (a < b) ? uint16_t(0) : uint16_t(a - b);
 }
 
 template <typename T> constexpr T max(T a, T b) { return (a > b) ? a : b; }
